renamed focus domain file name
